Split examples
This changeset split examples between small, simple, user-facing examples and longer, more complex, full-autotuning run benchmarks
