+++ CMakeLists.txt
@@ -258,8 +258,14 @@ endif()
 enable_testing()
 add_subdirectory(test)
 
-if (WITH_CAFFE2 AND WITH_CUDA)
+if (WITH_CUDA)
   add_subdirectory(examples)
 else()
-  message(STATUS "Not building examples, caffe2 or CUDA not available")
+  message(STATUS "Not building examples, CUDA not available")
+endif()
+
+if (WITH_CAFFE2 AND WITH_CUDA)
+  add_subdirectory(benchmarks)
+else()
+  message(STATUS "Not building benchmarks, caffe2 or CUDA not available")
 endif()
