+++ benchmarks/scripts/8.0.sh
