+++ benchmarks/scripts/9.0.sh
