+++ benchmarks/scripts/autotuner_parallel.sh
@@ -13,15 +13,15 @@ export TUNER_GPUS=${TUNER_GPUS:="0,1"}
 export GPU_NAME=$(nvidia-smi -L | head -n 1 | cut -d'(' -f 1 | cut -d':' -f 2 | sed "s/ //g")
 
 export TC_PREFIX=$(git rev-parse --show-toplevel)
-export PREFIX=${TC_PREFIX}/examples/results_$(date +%m%d%y)/${GPU_NAME}
-export LOG_DIR=${TC_PREFIX}/examples/results_$(date +%m%d%y)/${GPU_NAME}/logs/${SLURM_ARRAY_JOB_ID}-${SLURM_ARRAY_TASK_ID}
+export PREFIX=${TC_PREFIX}/benchmarks/results_$(date +%m%d%y)/${GPU_NAME}
+export LOG_DIR=${TC_PREFIX}/benchmarks/results_$(date +%m%d%y)/${GPU_NAME}/logs/${SLURM_ARRAY_JOB_ID}-${SLURM_ARRAY_TASK_ID}
 
 mkdir -p ${LOG_DIR}
 mkdir -p ${LOG_DIR}/autotuner
 chmod -R 777 ${LOG_DIR}
 
-cat ${TC_PREFIX}/examples/scripts/AUTOTUNER_COMMANDS | grep -v "\#" | head -n ${SLURM_ARRAY_TASK_ID} | tail -n 1 | xargs -i echo {} > ${LOG_DIR}/COMMAND
-cat ${TC_PREFIX}/examples/scripts/AUTOTUNER_COMMANDS | grep -v "\#" | head -n ${SLURM_ARRAY_TASK_ID} | tail -n 1 | xargs -i bash -c "{}"
+cat ${TC_PREFIX}/benchmarks/scripts/AUTOTUNER_COMMANDS | grep -v "\#" | head -n ${SLURM_ARRAY_TASK_ID} | tail -n 1 | xargs -i echo {} > ${LOG_DIR}/COMMAND
+cat ${TC_PREFIX}/benchmarks/scripts/AUTOTUNER_COMMANDS | grep -v "\#" | head -n ${SLURM_ARRAY_TASK_ID} | tail -n 1 | xargs -i bash -c "{}"
 
 # Run with:
-# sbatch --array=1-14 ./examples/scripts/autotuner_parallel.sh
+# sbatch --array=1-14 ./benchmarks/scripts/autotuner_parallel.sh
