+++ examples/tensordot.cc
